Added the stock service to fetch a particular stock price if present under registry

+++ backend/app/data_providers/yahoo_provider.py
@@ -17,7 +17,7 @@ class YahooMarketDataProvider(MarketDataProvider):
         "high": float(row["High"]),
         "low": float(row["Low"]),
         "close": float(row["Close"]),
-        "volume": float(row["Volume"])
+        "volume": int(row["Volume"])
       })
 
     return data
